:scroll: TypeScript clean-up of Frak and elimination of FrakTypes
