+++ src/global.d.ts
@@ -1,6 +1,6 @@
 import 'reactn';
 import {DrugLogRecord, MedicineRecord, ResidentRecord} from "./types/RecordTypes";
-import {ProvidersType} from "./types/FrakTypes";
+import {ProviderTypes} from "./types/ProviderTypes";
 
 declare module 'reactn/default' {
 
@@ -29,7 +29,7 @@ declare module 'reactn/default' {
         value: string;
         activeResident: ResidentRecord | null;
         development: boolean;
-        providers: ProvidersType;
+        providers: ProviderTypes.Providers;
         drugLogList: DrugLogRecord[] | null;
         medicineList: MedicineRecord[] | null;
         otcList: MedicineRecord[] | null;
